FIX: Rust compile errors — cast chain, u8 vs f64 type mismatches, unused imports
3 compile errors + 2 warnings fixed:

1. end_of_season.rs:1078 — Invalid 'as i64 as f64' cast chain:
   Rust doesn't allow 'as i64 as f64' without parentheses. Rewrote to
   keep the expression as f64 throughout and only cast to i64 at the end.

2. ai_hiring.rs:363 — manager.rating() returns u8, not f64:
   Changed sacked_candidate type from Option<(String, String, f64)> to
   Option<(String, String, u8)>. Changed max_by with partial_cmp to
   max_by_key (u8 implements Ord).

3. ai_hiring.rs:486 — Same u8 vs f64 mismatch in poaching candidates:
   Changed candidates Vec type from Vec<(..., f64)> to Vec<(..., u8)>.
   Changed sort_by with partial_cmp to sort_by with cmp.

4. live_match_manager.rs:6-7 — Unused import warnings:
   domain_to_engine_role and domain_to_engine_tactics re-exports are
   no longer consumed after C10's delegation refactor. Added
   #[allow(unused_imports)] to suppress the warning while keeping the
   re-exports available for future use.

These were the only compile errors in the V99.10 codebase.

+++ src-tauri/crates/ofm_core/src/end_of_season.rs
@@ -1075,7 +1075,9 @@ pub fn process_end_of_season(game: &mut Game) -> EndOfSeasonSummary {
                     .min(wage_cap)
                     .max(0);
                 // Transfer budget: 20% of finance × board_type multiplier.
-                team.transfer_budget = (((team.finance as f64 * 0.20) as i64 as f64)
+                // V99.10 fix: Removed invalid `as i64 as f64` cast chain —
+                // keep it as f64 throughout the multiplication, then cast to i64 at the end.
+                team.transfer_budget = ((team.finance as f64 * 0.20)
                     * board_transfer_mult) as i64
                     .max(0);
             }
